requirements.txt
Django>=2.2,<2.2.7
djangorestframework>=3.9.0,<3.10.3
psycopg2>=2.7.5,<2.8.4
#psycopg2-binary=2.8.4
Pillow==6.2.1

flake8>:3.6.0,<3.7.9
django-jet==1.0.8
drf-yasg==1.17.1